Allow nested events in buffered event bus
Because we want to be able to let events happen as a consequence of a certain event, the buffered event bus needs to make that possible.

Trying to flush the bus when already flushing, will no longer cause an Exception as the bus will be flushed if necessary after the current buffer has been flushed.