Revert "Merge pull request #160 from SURFnet/feature/registrants-should-not-be-able-to-vet-outside-institution"
This reverts commit a1543deca6a1f989b12eca2c119644b3fa951a6a, reversing
changes made to 4b0c8a62ef1182754d887529d2f98ca4f20ae291.

In the Identity aggregate, it is not known whether a certain identity is an SRAA or not. If he is an SRAA, the registrationAuthority property is set to null -- contrary to what one might assume.
That is why these checks currently do not work.

In the future, we need to find another way to determine in the Identity aggregate whether the one doing the vetting is a registration authority and whether he is an SRAA.
See: https://www.pivotaltracker.com/story/show/135335407

Reverting this means that, in theory, any Identity can vet a second factor regardless of whether they are within the same institution.
However, these checks are already done by the front-ends: RA and SS.