Validate institution names case-insensitively
The database is collated in a case-insensitive manner. As far as the database is concerned "Institution-a" == "institution-a".

This is why, when replacing an institution on the whitelist with the same institution name but with different casing,
the configured_institution is not changed: the InstitutionConfigurationProcessor detects no changes.
That is how you can have different casing as a configured_institution and as a whitelist_entry.

This goes wrong when trying to validate an institution configuration command as we check whether or not the
institution to configure is actually listed as a configured_institution.

This check was case-sensitive while its behavior should mirror the database, which works case-insensitive.