Allow registration of multiple tokens
A config parameter can be set to indicate the maximum number of tokens
allowed for each identity. The IdentityCommandHandler sets this value
on the identity instance. This in turn allows the identity instance to
validate whether or not the requested additional token is allowed or
not.

A sensible default was set in the parameters.yml.dist file.