Add reminder email command for verified tokens
Sending of reminders is implemented by utilizing the Symfony console
command feature. The command can be run with two optional parameters
being:

 1. The date the reminders should be processed of
 2. Dry run flag

The command is loaded on the Middleware bundle, and can be run using the
app/console middle:cron:email command.

The actual implementation is encapsulated in the
VerifiedSecondFactorReminderService. Which is handled using the newly
added ReminderEmailCommandHandler. The service in turn depends on a
mailer service to actually send the mail messages.