Do not validate SecondFactorTypes in Command
The Symfony validator was testing if the configured second factor tokens
where of the correct type when creating the
ReconfigureInstitutionConfigurationOptionsCommand. This was done by
calling a non existent static method on a class that was refactored
recently.

It's not possible to use the getAvailableSecondFactorTypes method that
was moved to the SecondFactorTypeService (stepup bundle) as this method
can not be called statically. It depends on second factor token config
that is injected on said Service.

The assertion can be removed as the validation will be performed at a
later stage when handling the command.