Add configuration option to toggle email verification
A new institution configuration options 'verify_email' allows
administrators to enable or disable the verification of email
addresses for users of a specific institution when registering a new
token.

The default behaviour of stepup is to require email verification. This
applies to all institutions without configuration, and institutions
with existing configurations prior to this commit. Starting from this
commit, institution configurations must explicitly enable or disable
the 'very_email' option.