Skip email verification if disabled for the institution
The default behaviour of stepup is to require email verification. When
the user has proven possession of a second factor, selfservice
triggers the 'possession proven' command, which is handled by
middleware as follows:

 - an unverified second factor is created
 - a verification email is sent
 -> user clicks on the verification link in the email
 -> the unverified second factor transitions to verified second factor
 -> an email containing the registration code is sent

This transition is fundamental to stepup and not easily changed. To
allow for users to skip email verification without having conditional
logic in many places, a second factor still goes trough the same
transition when email verification is disabled. When possession of a
second factor is proven the following happens in the same request:

 - an unverified second factor is created
 - no verification email is sent
 -> the unverified second factor transitions to verified second factor
 -> an email containing the registration code is sent