Require email verification by default
While not intended, email verification was disabled for existing
institutions after upgrading to 2.7.0. Enabling email verification
for those institutions was tricky because the user first has to
'disable' it before it could be enabled again.

This commit fixes the problem by defaulting the column to true,
matching the expectation of underlying code.