Ensure allowed_second_factor projection is updated
When removing allowed second factors from an existing institution config
, in the previous situation, a Doctrine unique constraint violation is
raised. This because the allowed_second_factor projection is not flushed
before adding existing entries.

The solution to this problem is to always clear the
allowed_second_factor for the institution that is currently updated.
After clearing the data, the new entries are persisted.

For more details about the problem please read the bug report in Pivotal
https://www.pivotaltracker.com/story/show/156466792