Read # of tokens setting from service
Now that the number of tokens per identity is no longer a global yaml
configuration option, the setting is read from the institution
configurations options service. The service is also injected with the
application global setting. This will be used as the fallback setting
when no specific value is configured for an institution.