Remove possibly risky update query
While fixing a default value bug[1] for the email verification
configuration option, the additional UPDATE query in the migration
(Version20180330094402) was not removed. This is risky and not necessary
as the default value is set in the ALTER TABLE and the column is not
nullable.

This issue was identified during review of the changes described in the
associated PR. The fix for this migration was boyscouted.

See:
1. https://github.com/OpenConext/Stepup-Middleware/pull/215