Update NewInstitutionConfigurationCreatedEvent
When a new institution is created, it will be set with default settings
for the FGA options. These being null, not set.

This required the addition of the getDefault method on the Option
classes.

As an addition a small inconsistency was fixed. getSelectRaaOptions was
renamed to getSelectRaaOption (minus the s)

Finally the InstitutionConfigurationTest was updated to use the new
options. And a test was added that tests specifically if the sorting of
the fga options does not affect the aggregate state.