Add InstitutionAuthorizationContextInterface
The InstitutionAuthorizationContextInterface was created, describing the
way an authorization interface should look like.

One of the characteristics is that role requirements are set on the
context. Which are packaged in the InstitutionRoleSet.

The InstitutionRoleSet is added for the sake of being able to work with
the interface, the definitive implementation and tests will be taken
care of in successive commits.

See: https://www.pivotaltracker.com/story/show/160983081