Distinguish institution filtering & authorization
When filtering the second factor token entries, we showed entries for
the institutions the authenticated user (actorInstitution) is allowed
to see.

Filtering on institutions is also a requirement. In order to achieve
this goal, the actorInstitution is introduced on the search query.

The Repository now uses the actorInstitution for determination of the
authContext and the institution for filtering.