Ensure RA filter applied when searching verif. SF
RA's are able to vet second factor tokens, in order to be able
to find the verified second factor, the RA searches for the token by
activation code. Before this fix, the RA would not find the token as the
controller was configured to require the RAA role in order to find a
token.