Use configured institutions for institution lists
The used institution listing are only added when a user is created.
Therefore it's impossible to select an institution as SRAA to see
institution configurations and to perform (S)RAA actions on
institutions if a user from that institution hasn't logged in before.
This will prevent the usage of 'virtual' institutions because no user
will be able to login with that institution as sho.

The InstitutionListing- Repository and Projector are now only used
to support a migration `Version20160719090052.php`. This could
ultimately be dropped.