Fix institution accreditation check
When the old institution configuration is used the check to validate
if a users institution may change it's own institution will fail.
Because the InstitutionConfiguration aggregate doesn't have theevent
SelectRaaOptionChanged event the list with selectRaa institutions
will be null and therefore the method on this collection could not be
called.