Add 'self_vet' option to institution configuration
This option was added throughout the application. The setting will be
saved to the event store, but is also projected to the institution
configuration projection. This way the setting can be retrieved via the
Middelware management api. At this location the option is also evaluated
to store it for both new and existing institutions. For institutions
without the configuration option, the default 'false' is configured.

See: https://www.pivotaltracker.com/story/show/176714989