Create token migration command
A Token command is added to be able to move second factors
from a source to a target identity. Pre conditions are required
to allow this move. The command is idempotent and it is possible
to create an idenity on thge fly if needed but a institution
name and a email address should be supplied.