Correct the return value according to data type
The options are tinyint fields in the database. The boolean that is used
in the application to represent this value was used to serialize the
database value.

This raises an error when the DBMS is running in strict mode. As it will
not cast the boolean to an integer value.

By casting the boolean to an integer value, this problem is resolved.