Implement the deprovision logic
Deprovisioning is simply utilizing the existing right to be forgotten
command. This command is now also allowed to be executed by the
Deprovision (lifecycle) user.

The output of this api call returns the user data that was removed by
first calling the obtain data command. And then calling the silent
deprovision command.

Next task is to correctly format the output data.