Introduce self-asserted tokens to Identity
Here we for now only verify if the self-asserted tokens feature can
be applied or not. In other words; the pre conditions are formulated on
the Identity aggregate on when the identity is allowed to self-assert a
token.

Note that no verification on the institution configuration is applied
here. This because that is not part of this bounded context. That check
will be performed on the command handler that interacts with the
aggregate.