Create the AuthorizationService
This service can be used to perform in depth authorization decisions,
other than the fine grained authorization decision which are covered in
the neighboring service.

Opposed to the FGA authorizations which are always applied to the scope/context
of a query or command. The AuthorizationService can also be used via the
Middleware API by the SS or RA applications to indicate if certain
actions are allowed for a given identity.