Apply SAT option changed event on aggregate
That way the state of the aggregate is up to date when using with the
configuration. This bug was found during functional testing the recovey token
feature. Switching the option on/off did not affect the fact that you
should not be able to add a rec-token when the option is disabeld.