Introduce a helper to deal with Secrets
The RecoverTokenSecretHelper is a very simple helper class that calls
the hash method on an unhashed hash VO. It then returns the hashed
counterpart to the callee. Why was this helper introduced? Mainly to
aid testability. Having a service that can be mocked makes it possible to
make predictable assertions on the hashed secrets. Given our implementation
of the bcrypt, a different hash was generated on every test-run. This because
the salt is randomly generated for every hash operation.

This commit includes both the creation of the service (helper) and the
integration ito the command handler.

Having the helper to hash in the command handler was a tradeoff. Ideally
we'd like this action performed in the Identity aggregate. However, the action
is performed within the domain, so we're good on that front.