Deny SAT when both RT and SAT SF are lost
If a user looses both their activated token and their recovery method they
cannot activate a token themselves. Allowing this would enable an attack
whereby the attacker with only the first factor can delete and then add a new
token to the account, upgrading it to a 2FA account. So getting out of this
situation must involve an external verification step.

To do this, the user has to follow the normal service desk vetting process.
I.e. register a token and then select the service desk activation method.
The result is that the user will then have a normal loa2 or loa3 token.
This is a higher level than the user had before, but this will work just fine.
Using this process enables us to implement this scenario using the existing
Stepup functionality.

The check missing in the existing authz check was to verify that the
user does have an active recovery token.

https://www.pivotaltracker.com/story/show/181934196