Process VettingTypeHintCommand
The vetting type hints are now processed and saved in the
VettingTypeHintsSavedEvent. This event is not yet projected, but that's
nex on the agenda.

The hints themselves are saved in a collection that keys the translation
locales with the hint texts that where received from the command body.
The locales that are actually processed are the ones configured under
the `locales` config in parameters.yaml. Take note that they should (but
no must be) equal to those of the RA. If the RA app configures a
different set of locales. Only the ones known to Middleware are
processed.A warning is logged for each language that is lost that way.

See: https://www.pivotaltracker.com/story/show/182129438