Update translations
In a broader sense than the usual additions to the translation files:
here the config was changed slightly too. Extracted translations landed
in the wrong folder and the jms_translation config files where added in
abbundance without adding environement specific adjustments to the
default settings. This was updated by creating a base jms_translation
file.