Wrap address and name into an Address
The to and from methods of the `TemplatedEmail` ask for addresses as its
parameter. They are resolved quite magically. Setting a string with a
plain mail address is possible, adding multiple parameters is cool too.

The issue we ran into was that the addres param was followed by a common
name string. Both where resoved as being mail addresses. Causing the
application to blow up on the common name. The solution is to provide
Address objects. An address consists of an address and a name.

See: https://www.pivotaltracker.com/n/projects/1163646/stories/183362518
See: https://symfony.com/doc/4.4/mailer.html#email-addresses