Refrain from injecting old mailer
Some of our other dependencies stil rely on swiftmailer. And hence we
have the mailer in the di container. On my previous (not to thorough, ahem)
review I did not notice this discrepency. And the mailers would keep
working as the interface change between the Swift_mailer and the SF
Mailer is not changed on the methods we use.

Except for the ones where we passed the (email, name) construction
without wrapping them in an Address (see previous PR#377).

Quite the pickle!

See: #377
See: https://www.pivotaltracker.com/n/projects/1163646/stories/183362518