Remove old swiftmailer config files
They cause the application to try and load the swiftmailer bundle. This
was placed on the original PR, but was never reworked. And I in te end
just hit the merge button

(╯°□°)╯︵ ┻━┻

I merged the old config with the new mailer config settings as good as
possible. The one change required was to disable delivery in the test
env. This was achieved by setting the DSN to null://null. As suggested
in: https://symfony.com/doc/4.4/mailer.html#disabling-delivery

See: #352 #377 #378