Create the sso_on_2fa feature toggle
In order to enable/disable the sso on 2fo feature, a feature toggle per
institution is required. Allowing us to enable the feature on a per
institution basis.

The ammount of work required to achieve this task is vast to say the
least. In short:

1. The existing NewInstitutionConfigurationCreatedEvent was updated with
   the new option. And a new SsoOn2faOptionChangedEvent was created.
2. Support for the new SsoOn2faOption was added to the Command an
   CommandHandler.
3. Database storage (in the insittution_configuration projection) was
   realised.
4. Tests where updated.

See: https://www.pivotaltracker.com/story/show/183402518