Change 'default value' behavior
Previously, not configuring an inst-conf option would result in it not
being evaluated in the command handling.

That was not the intended behavior though. When the configuration is not
present in the POSTed payload, the value should revert to the default
value.

This is implemented in the command handler for the three boolean config options
that are optional