Add identity_vetted column to Second Factor entity
As per point 1 of https://www.pivotaltracker.com/story/show/184191063

It identifies if a RA vetted vetting type was used to vet the SF token.
The on-premise and self-vet methods are considered identity-vetted
vetting types. The upcoming self-asserted tokens are not.

By reducing the vetting types into this new boolean value, we can set
the default value of this column correctly. This because all vetted second
factors are currently either ra-vetted or self-vetted.