Update SAT projections
The SAT migration was updated to ensure it is safe to use in a rolling
update scenario. In other words. All columns are populated with sensible
data so both the old and new versions can function correctly with the
SAT database schema.

See task 3 of: https://www.pivotaltracker.com/story/show/184191063