Support self-vetting of self-asserted tokens
When an Identity is in possession of a vetted self-asserted token. It
should be allowed to self-vet successive tokens using that SAT. But only
when all tokens are of the SAT type. Once an Identity is in possession
of an idenity vetted token (vetted on-premise), all following self
vetting tokens must be done with the identity vetted token.

For details see: https://www.pivotaltracker.com/story/show/184292087
And: https://github.com/OpenConext/Stepup-SelfService/pull/284