Create GW Inst Config if not present in projection
When the Gateway institution configuration is updated because the
sso on 2fa config changed for that institution. The projection would
only be updated if the institution already exists in the projection. If
the projection did not yet have the institution, then the projection was
not updated.

This fix now creates the row if the projection is not yet present.