Determine authorizations based on RA listing
The institution configuration FGA auth rules are evaluated to specify
which institutions the RA(A) is active at. But these rules were not
matched against the actual rules accredited to the RA(A). This made it
possible that a RA would be granted RAA rights even tho it had only RA
rights for his/her institution(s)

By evaluating the RA listing authorizations alongside the insittutional
authz, we now have a realistic vision of the roles of the RA(A)