Make expiration date in mail match the one displayed in selfservice.
Selfservice renders the date without ->endOfDay(). Adding the endOfDay
makes the date wrap to the next day when the timezone modifier is applied.

More intricate timezone aware handling of this date is thinkable,
but this change at least brings both interfaces to use the same algorithm
so the output will be the same.