Address level 4 PHPStan issues
Doctrine extention was added to the Stan setup. Preventing unused
field warnings.

Three warnings have been ignored. They are regarding a hidden use of
three RegistraoinAuthority fields. Static analysis does not evaluate
that they are written when persisted to db storage

The rest of the fixes are mainly removal of unused code fragments. Or
removal of redundant sections of logical expressions.