Plain package update
In test/development mode, don't flood the console with queries.

Update mariaDb version to match production.

In the devbox, we want deprecations to fail the pipeline. In Github, we don't want that, as dependabot updates or small package udates should not cause the build to fail. We will handle the deprecations once a developer works on the project.

 Add a separate deprecation.log in development mode.