AllowDynamicProperties on RegisterSelfAssertedSecondFactorCommand and UpdateIdentityCommand
Prior to this change, `Deprecated: Creation of dynamic property` was logged when running the behat tests in devconf. This can be caused when properties are set in old events, that are later removed from the class.

We could check in the CommandValueResolver whether a field exists before setting it, but that carries more risk, in case the properties might be still read dynamically somewhere.